warn about implicit floating point to integer conversions
This got disabled when I had to disable '-Wdouble-promotion' due to it
being practically unavoidable with variadic functions.

But while '-Wdouble-promotion' wasn't practical, I should have kept
'-Wfloat-conversion' around.

Yes, it requires a few manual casts for out hacky 'pow()' usage that
I'll need to fix for the actual real-time microcontroller case, but
those are good to have anyway, not the unavoidable issue with the
calling conventions.

That makes a couple of bugs in the growlingbass implementation obvious.

+++ tube.h
@@ -89,7 +89,7 @@ static float tube_step(float in)
 {
 	in *= tube.boost;
 	if (in+1 > 0)
-		in = pow(in + 1, 1.5)-1;
+		in = (float)pow(in + 1, 1.5)-1;
 	else
 		in = -1;
 	in *= tube.volume;
